+++ tasks/options/jshint.js
@@ -1,0 +1,13 @@
+/**
+ * JSHint taks configuration
+ */
+'use strict';
+
+var config = require('../config');
+
+module.exports = {
+	all: config.jsHintFiles,
+	options: {
+		jshintrc: '.jshintrc'
+	}
+};